move utils into seperate file
